# application/ui/src/api/openapi-spec.json
{"openapi":"3.1.0","info":{"title":"Geti Action","description":"Geti Action is a framework to train robots.It allows the user to create datasets, models and the run inference ","version":"0.1.0"},"paths":{"/api/projects":{"get":{"summary":"List Projects","description":"Fetch all projects","operationId":"list_projects_api_projects_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Project"},"type":"array","title":"Response List Projects Api Projects Get"}}}}}},"put":{"summary":"Create Project","description":"Create a new project","operationId":"create_project_api_projects_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/projects/{project_id}/import_dataset":{"put":{"summary":"Impport Dataset","description":"Sets the project from a dataset, only available when config has not been set yet","operationId":"impport_dataset_api_projects__project_id__import_dataset_put","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeRobotDataset"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/projects/{project_id}":{"delete":{"summary":"Delete Project","description":"Delete a project","operationId":"delete_project_api_projects__project_id__delete","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/projects/{id}":{"get":{"summary":"Get Project","description":"Get project by id","operationId":"get_project_api_projects__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/projects/{project_id}/datasets/{repo}/{id}":{"get":{"summary":"Get Dataset Of Project","description":"Get dataset of project by id","operationId":"get_dataset_of_project_api_projects__project_id__datasets__repo___id__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}},{"name":"repo","in":"path","required":true,"schema":{"type":"string","title":"Repo"}},{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeRobotDataset"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/hardware/cameras":{"get":{"summary":"Get Cameras","description":"Get all cameras","operationId":"get_cameras_api_hardware_cameras_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Camera"},"type":"array","title":"Response Get Cameras Api Hardware Cameras Get"}}}}}}},"/api/hardware/robots":{"get":{"summary":"Get Robots","description":"Get all connected Robots","operationId":"get_robots_api_hardware_robots_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/RobotPortInfo"},"type":"array","title":"Response Get Robots Api Hardware Robots Get"}}}}}}},"/api/hardware/calibrations":{"get":{"summary":"Get Lerobot Calibrations","description":"Get calibrations known to huggingface leRobot","operationId":"get_lerobot_calibrations_api_hardware_calibrations_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CalibrationConfig"},"type":"array","title":"Response Get Lerobot Calibrations Api Hardware Calibrations Get"}}}}}}},"/api/hardware/identify":{"put":{"summary":"Identify Robot","description":"Visually identify the robot by moving given joint on robot","operationId":"identify_robot_api_hardware_identify_put","parameters":[{"name":"joint","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Joint"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RobotPortInfo"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/hardware/camera_feed":{"get":{"summary":"Get Camera Feed","description":"Get a streaming response from the camera","operationId":"get_camera_feed_api_hardware_camera_feed_get","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"string","title":"Id"}},{"name":"type","in":"query","required":true,"schema":{"enum":["RealSense","OpenCV"],"type":"string","title":"Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cameras/offer/camera":{"post":{"summary":"Offer Camera","description":"Create a WebRTC offer","operationId":"offer_camera_api_cameras_offer_camera_post","parameters":[{"name":"camera","in":"query","required":true,"schema":{"type":"string","title":"Camera"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Offer"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Answer"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/dataset/lerobot_datasets":{"get":{"summary":"List Leorobot Datasets","description":"Get all local lerobot datasets from huggingface cache","operationId":"list_leorobot_datasets_api_dataset_lerobot_datasets_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/LeRobotDataset"},"type":"array","title":"Response List Leorobot Datasets Api Dataset Lerobot Datasets Get"}}}}}}},"/api/dataset":{"get":{"summary":"List Datasets","operationId":"list_datasets_api_dataset_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Dataset"},"type":"array","title":"Response List Datasets Api Dataset Get"}}}}}}}},"components":{"schemas":{"Answer":{"properties":{"sdp":{"type":"string","title":"Sdp"},"type":{"type":"string","title":"Type"}},"type":"object","required":["sdp","type"],"title":"Answer"},"CalibrationConfig":{"properties":{"id":{"type":"string","title":"Id"},"path":{"type":"string","title":"Path"},"robot_type":{"type":"string","enum":["teleoperator","robot"],"title":"Robot Type"}},"type":"object","required":["id","path","robot_type"],"title":"CalibrationConfig"},"Camera":{"properties":{"name":{"type":"string","title":"Name","description":"Camera name"},"id":{"type":"string","title":"Id","description":"Either serial id for  RealSense or port for OpenCV"},"type":{"type":"string","enum":["RealSense","OpenCV"],"title":"Type"},"default_stream_profile":{"$ref":"#/components/schemas/CameraProfile"}},"type":"object","required":["name","id","type","default_stream_profile"],"title":"Camera"},"CameraProfile":{"properties":{"width":{"type":"integer","title":"Width"},"height":{"type":"integer","title":"Height"},"fps":{"type":"integer","title":"Fps"}},"type":"object","required":["width","height","fps"],"title":"CameraProfile"},"Dataset":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name","default":"Default Name"},"path":{"type":"string","title":"Path"}},"type":"object","required":["path"],"title":"Dataset","example":{"name":"Collect blocks","path":"/some/path/to/dataset"}},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LeRobotDataset":{"properties":{"root":{"type":"string","title":"Root"},"repo_id":{"type":"string","title":"Repo Id"},"total_episodes":{"type":"integer","title":"Total Episodes"},"total_frames":{"type":"integer","title":"Total Frames"},"features":{"items":{"type":"string"},"type":"array","title":"Features"},"fps":{"type":"integer","title":"Fps"}},"type":"object","required":["root","repo_id","total_episodes","total_frames","features","fps"],"title":"LeRobotDataset"},"Offer":{"properties":{"sdp":{"type":"string","title":"Sdp"},"type":{"type":"string","title":"Type"},"webrtc_id":{"type":"string","title":"Webrtc Id"}},"type":"object","required":["sdp","type","webrtc_id"],"title":"Offer"},"Project":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name","default":"Default Name"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"},"config":{"anyOf":[{"$ref":"#/components/schemas/ProjectConfig"},{"type":"null"}],"description":"Project config"},"datasets":{"items":{"$ref":"#/components/schemas/Dataset"},"type":"array","title":"Datasets","description":"Datasets","default":[]}},"type":"object","title":"Project","example":{"datasets":[{"id":"fec4a691-76ee-4f66-8dea-aad3110e16d6","name":"Collect blocks","path":"/some/path/to/dataset"}],"id":"7b073838-99d3-42ff-9018-4e901eb047fc","name":"SO101 Teleoperation","updated_at":"2021-06-29T16:24:30.928000+00:00"}},"ProjectConfig":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"fps":{"type":"integer","title":"Fps","description":"Recording FPS for datasets","default":30}},"type":"object","title":"ProjectConfig","example":{"fps":"30"}},"RobotPortInfo":{"properties":{"port":{"type":"string","title":"Port"},"serial_id":{"type":"string","title":"Serial Id"},"device_name":{"type":"string","title":"Device Name"}},"type":"object","required":["port","serial_id","device_name"],"title":"RobotPortInfo"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}